requirements.txt
openpyxl==3.1.2
pandas==2.0.3
Jinja2==3.1.2
python-dotenv==1.0.0